batch upload scores

+++ studentManagement/src/main/resources/db/schema.sql
@@ -24,7 +24,7 @@ create table if not exists  stu_score (
     username varchar(50) NOT NULL COMMENT 'user name',
 	academic_year varchar(50) NOT NULL COMMENT '学年',
 	semester int NOT NULL COMMENT '学年',
-	score int NOT NULL COMMENT '成绩',
+	score float NOT NULL COMMENT '成绩',
     course varchar(50) NOT NULL COMMENT '课程',
     credit float NOT NULL COMMENT '学分',
     course_nature int NOT NULL COMMENT '课程性值:必修/选修/体育',
